[TNT-114] refactor: 회원가입 완료 화면 파일명 수정
